feat: bun and nodejs parity tests
