+++ tests/parity/node-consumer.ts
@@ -1,10 +1,11 @@
 import { setUpWorkers } from './consumer';
-import { getBackend, logEvent } from './utils';
+import { getBackend } from './utils/node';
+import { logEvent } from './utils/shared';
 
 async function main() {
   const backend = getBackend();
 
-  setUpWorkers(backend.options, backend.factory);
+  await setUpWorkers(backend.options, backend.factory);
 
   // Consumer creates workers first, then sends a ready event
   logEvent('ready');
