+++ tests/parity/utils/node.ts
@@ -1,13 +1,11 @@
-import { readFile } from 'node:fs/promises';
 import {
   BackendFactory,
-  createBunRedisClient,
   createPostgresBackend,
   createRedisBackend,
   PostgresQueueBackend,
   QueueOptions,
   RedisQueueBackend,
-} from '../../src';
+} from '../../../src';
 
 interface ParityBackend {
   factory: BackendFactory<PostgresQueueBackend | RedisQueueBackend>;
@@ -41,25 +39,4 @@ export function getBackend(): ParityBackend {
   }
 
   throw new Error(`Invalid PARITY_BACKEND value ${typeof backend}(${backend})`);
-}
-
-export function logEvent(event_type: string, data?: any) {
-  const run_id = process.env.PARITY_RUN_ID;
-  console.log(JSON.stringify({ type: event_type, run_id, data }));
-}
-
-export async function readDefinitons<T>(): Promise<T[]> {
-  const content = await readFile('./parity/definitions.json', 'utf-8');
-  return JSON.parse(content);
-}
-
-export async function getBunRedisBackendConnection() {
-  // Bun is not guaranteed to be there on every run
-  const { RedisClient } = await import('bun');
-  const port = process.env.PARITY_BACKEND_PORT;
-  const rawClient = new RedisClient(`redis://localhost:${port}`);
-
-  const connection = createBunRedisClient(rawClient);
-
-  return connection;
 }
